requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
biopython==1.84
blurhash==1.1.4
boto3==1.35.83
botocore==1.35.83
Brotli==1.1.0
celery==5.4.0
certifi==2024.12.14
cffi @ file:///croot/cffi_1726856441404/work
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
conda-pack==0.8.1
contourpy==1.3.1
cryptography @ file:///croot/cryptography_1732129644106/work
cycler==0.12.1
decorator==5.1.1
deptry==0.21.1
Django==5.1.4
django-bootstrap5==24.3
fonttools==4.55.3
freetype-py==2.3.0
greenlet @ file:///croot/greenlet_1733860072744/work
gunicorn==23.0.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
jmespath==1.0.1
kiwisolver==1.4.7
kombu==5.4.1
Mastodon.py==1.8.1
matplotlib==3.10.0
munkres==1.1.4
numpy==2.2.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pycairo==1.20.0
pycparser==2.22
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-magic==0.4.27
pytz @ file:///croot/pytz_1713974312559/work
rdkit==2024.3.6
rdkit-pypi==2022.9.5
redis==5.2.1
reportlab==4.2.5
requests==2.32.3
requirements-parser==0.11.0
rlPyCairo==0.2.0
s3transfer==0.10.4
simplejson @ file:///croot/simplejson_1723144746668/work
six==1.17.0
SQLAlchemy==2.0.36
sqlparse==0.5.2
tomli==2.2.1
types-setuptools==75.6.0.20241126
typing_extensions==4.12.2
tzdata==2024.2
unicodedata2==15.1.0
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
win-inet-pton==1.1.0
zipp==3.21.0
zstandard==0.23.0
